Fix Interlacing issue with CAN Analog Axis 5-6
There is a bug in the CanAnalog.cpp buffer for the axis 5 & 6. When the first 4 axis are being read axis 5 & 6 will drop to zero and will not retain their value until the next read. This change fixes the issue and is tested and working. Recommend adding this change to CanAnalog.cpp.